```
from asv_bench.benchmarks.groupby import *
self = Apply()
self.setup(4)

%timeit self.time_copy_overhead_single_col(4)
213 ms ± 10.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
161 ms ± 2.84 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```